3.1.1 (gallery) (#136)
* [skip] fix maven publication

* (1.20.1) fix chatWidget appearance

* (1.8.9, 1.16_combat-6) add option to disable inventory potion offset

* begin gallery implementation

* add loading animations

* add automatic reloading on file updates

* port to the other versions

* set mojang auth to only use http1.1

* fix various bugs

* fix BedwarsPlayerStats parameter mismatch

* fix more bugs

* fix another batch of bugs

* add options to display durability on the ArmorHud

* add secondary key to freelook

* add option to allow friends to access uploaded images in gallery

* update API user agent to include game version

- bump config library version to fix color option saving

* fix CoordsHud bugs, add biome option

* fix more armorhud bugs

* (1.21.1) fix crash when using minimal item physics

* some bug fixes

* refactor some auth code again

* (bedwars module) add a useful message for unknown bed break messages

* fix another bug

* fix more bugs

- (1.21.1) remove custom sky module

* (1.8.9) fix more custom sky-related bugs

* skip skies with non-existent textures

* add new bedwars messages

- fix auth again

* fix screenshot actions only affecting the most recent screenshot

* Update fr_fr.json

* Update fr_fr.json

* (1.8.9) fix flashy chat animation

* best-effort name formatting for biome locations

- (1.20.1) fix nametag shadows

* account for embed changes

* fix back button position on DownloadImageScreen

- fix TntTime for 1.21.4, rendering independent of other name tag attachments
- make PingHud always ping the server

* (1.21.4) lower TntTime display

* adjust text field length

* fix another few bugs

* improve error handling for image upload failures

* (1.8.9) fix ChatHud color/transparency issues

* fix API-related buttons not appearing correctly on the TitleScreen

* (1.8.9) "fix" raw mouse input option

- by delegating to glfw since we now depend on it anyway

* (1.21.4) fix hitcolor having red &  blue channels swapped

* (1.8.9) fix retrieving the glfw window handle

* (1.20.1, 1.21.1) fix nametag-related bugs

* (1.21.1, 1.21.4) fix hud color differences

- (1.20.1) fix badge rendering

* fix oversight

* bump version

* fix more bugs

- some of which were only introduced in the previous commits

* small badge rendering tweaks

- fix race condition on API login

* (1.20.1) fix nametag shadow

* fix name shadow rendering while sneaking

---------

+++ 1.21.4/src/main/java/io/github/axolotlclient/modules/hud/gui/hud/KeystrokeHud.java
@@ -295,7 +295,7 @@ public class KeystrokeHud extends TextHudEntry {
 		protected final Rectangle bounds;
 		private final int animTime = 100;
 		protected DrawPosition offset;
-		private float start = -1;
+		private long start = -1;
 		private boolean wasPressed = false;
 
 		public Keystroke(Rectangle bounds, DrawPosition offset, KeyMapping key, KeystrokeRenderer render) {
@@ -311,7 +311,7 @@ public class KeystrokeHud extends TextHudEntry {
 		}
 
 		private float getPercentPressed() {
-			return start == -1 ? 1 : Mth.clamp((Util.getMillis() - start) / animTime, 0, 1);
+			return start == -1 ? 1 : Mth.clamp((float)(Util.getMillis() - start) / animTime, 0, 1);
 		}
 
 		public void render(GuiGraphics matrices) {
@@ -330,7 +330,7 @@ public class KeystrokeHud extends TextHudEntry {
 			if (outline.get()) {
 				outlineRect(matrices, rect, getOutlineColor());
 			}
-			if ((Util.getMillis() - start) / animTime >= 1) {
+			if ((float)(Util.getMillis() - start) / animTime >= 1) {
 				start = -1;
 			}
 			wasPressed = key.isDown();
